refactor(courtboard): extract WaitingList component
- Create src/courtboard/components/Icons.jsx with icon components
  (Users, TennisBall, Calendar, AlertCircle)
- Extract WaitingList to components/WaitingList.jsx
- Pass maxWaitingDisplay as prop instead of using TENNIS_CONFIG directly
- Remove inline definitions from main.jsx
